requirements.txt
requests
python-dotenv
# configparser
zammad_py
google-adk